requirements/requirements_pytorch.txt
hydra-core==1.0.0
pytorch_lightning==0.8.4
torch==1.6.0
torchvision==0.7.0

matplotlib
numpy
pandas
Pillow
opencv-python
tqdm
scikit-learn
scipy
jupyterlab
pycocotools

#detection
albumentations
beautifulsoup4
ensemble-boxes
numba

#keypoint
Cython
imageio
munkres
